Refactor language structures
